[fix] fix auto_finish.
audo_finish should finish all projects in the year which is specified
by the adminStaff, however in previous commits, the project queryset
is (NOT_OVER or IS_DELAY) and IS_NOT_PAST, so it cant' work properly
on others years due to IS_NOT_PAST is always True for past years. In
this commit, I fix this by overwrite looking up query.